(MODULES-7240) Setting User For A Task Fails
Prior to this change, if a manifest specified a user and a password to
run a task, those settings were not respected. Puppet's commandline
logging would first claim that a task was create, but then throw an OLE
error.

This change ensures that a username and password can successfully be
specified for a task.